feat: using proxied api instead of calling btc node directly
